requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
djangorestframework==3.12.4
Pygments==2.10.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
requests==2.24.0
